+++ src/test/resources-noncompilable/com/puppycrawl/tools/checkstyle/checks/design/onetoplevelclass/package-info.java
@@ -20,4 +20,4 @@
 /**
  * Package containing input files for design checks.
  */
-package com.puppycrawl.tools.checkstyle.checks.design;
+package com.puppycrawl.tools.checkstyle.checks.design.onetoplevelclass;
