--- src/test/resources/com/puppycrawl/tools/checkstyle/checks/design/InputOneTopLevelEnum.java
@@ -1,5 +1,0 @@
-package com.puppycrawl.tools.checkstyle.checks.design;
-
-public enum InputOneTopLevelEnum {
-    VALUE1, VALUE2;
-}
