--- src/test/resources/com/puppycrawl/tools/checkstyle/checks/design/InputOneTopLevelInterface.java
@@ -1,5 +1,0 @@
-package com.puppycrawl.tools.checkstyle.checks.design;
-
-public interface InputOneTopLevelInterface {
-    int foo();
-}
