+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/design/onetoplevelclass/InputOneTopLevelClassDeclarationOrder.java
@@ -1,6 +1,6 @@
-package com.puppycrawl.tools.checkstyle.checks.design;
+package com.puppycrawl.tools.checkstyle.checks.design.onetoplevelclass;
 
-public class InputDeclarationOrder
+public class InputOneTopLevelClassDeclarationOrder
 {
     private class InnerClass
     {
