+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/design/onetoplevelclass/InputOneTopLevelEnum.java
@@ -1,0 +1,5 @@
+package com.puppycrawl.tools.checkstyle.checks.design.onetoplevelclass;
+
+public enum InputOneTopLevelEnum {
+    VALUE1, VALUE2;
+}
