+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/design/onetoplevelclass/InputOneTopLevelInterface.java
@@ -1,0 +1,5 @@
+package com.puppycrawl.tools.checkstyle.checks.design.onetoplevelclass;
+
+public interface InputOneTopLevelInterface {
+    int foo();
+}
